http endpoint provider refactorings
