+++ internal/rules/provider/httpendpoint/provider_registrar.go
@@ -13,13 +13,6 @@ import (
 	"github.com/dadrus/heimdall/internal/x/errorchain"
 )
 
-type httpEndpointProvider struct {
-	endpoint     endpoint.Endpoint
-	watchChanges bool
-	queue        event.RuleSetChangedEventQueue
-	logger       zerolog.Logger
-}
-
 type registrationArguments struct {
 	fx.In
 
@@ -64,34 +57,5 @@ func registerProvider(args registrationArguments, logger zerolog.Logger) error {
 
 	logger.Info().Str("_rule_provider_type", "http_endpoint").Msg("Rule provider configured.")
 
-	return nil
-}
-
-func newProvider(
-	endpoint endpoint.Endpoint,
-	watchChanges bool,
-	queue event.RuleSetChangedEventQueue,
-	logger zerolog.Logger,
-) (*httpEndpointProvider, error) {
-	if err := endpoint.Validate(); err != nil {
-		return nil, errorchain.
-			NewWithMessage(heimdall.ErrConfiguration,
-				"failed to validate http_endpoint rule provider endpoint configuration").
-			CausedBy(err)
-	}
-
-	return &httpEndpointProvider{
-		endpoint:     endpoint,
-		watchChanges: watchChanges,
-		queue:        queue,
-		logger:       logger,
-	}, nil
-}
-
-func (p *httpEndpointProvider) Start() error {
-	return nil
-}
-
-func (p *httpEndpointProvider) Stop() error {
 	return nil
 }
